chore(mailchimp): typescripted and moved addUserToAudience to mailchimp module
